backend/requirements.txt
# Python 3.10–3.13 compatible; avoid pydantic 2.6.x on 3.13 (pydantic-core has no wheel)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.25
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
firebase-admin>=6.4.0
httpx>=0.26.0
elevenlabs>=1.5.0
numpy>=1.26.0,<3.0.0
openai>=1.0.0
